refactor: 네비게이션 타입 정의를 navigation 폴더로 이동
- 기존 types/navigation.ts에서 정의한 RootStackParamList, AuthStackParamList를
  App.tsx 및 AuthStack.tsx 내부로 이동하여 분산 관리
- 타입 임포트 경로를 navigation 폴더 기준으로 수정
- CompositeNavigationProp 사용을 제거하고 단순한 타입으로 간소화
