--- frontend/src/types/navigation.ts
@@ -1,9 +1,0 @@
-export type RootStackParamList = {
-  Auth: undefined;
-  Home: undefined;
-};
-
-export type AuthStackParamList = {
-  Login: undefined;
-  SignUp: undefined;
-};
